+++ app/specific/Main.js
@@ -226,6 +226,7 @@ function Main_StartApp() {
                     PlayHLS_GetTokenResult: PlayHLS_GetTokenResult,
                     PlayHLS_PlayListUrlResult: PlayHLS_PlayListUrlResult,
                     PlayHLS_AdProbeResult: PlayHLS_AdProbeResult,
+                    Play_AdBlockReload: Play_AdBlockReload,
                     AddCode_AppTokenResult: AddCode_AppTokenResult,
                     Play_UpdateDurationDiv: Play_UpdateDurationDiv,
                     Screens_PlaybackTimeSetVodDuration: Screens_PlaybackTimeSetVodDuration